feat: install compiler, tsx, and tsconfig for gradual migratio
