+++ .gitignore
@@ -39,3 +39,5 @@ jspm_packages
 .idea
 
 *.db
+
+dist
