+++ package.json
@@ -2,10 +2,13 @@
   "name": "amk",
   "version": "0.2.0",
   "description": "no frills API server using express",
-  "main": "api.js",
+  "main": "dist/src/api.js",
+  "types": "dist/src/api.d.ts",
   "scripts": {
-    "start": "node --env-file=.env app.mjs",
-    "test": "c8 node --env-file=.env.test --test --test-concurrency=1",
+    "build": "tsc",
+    "start": "node --env-file=.env dist/app.js",
+    "dev": "tsx --env-file=.env app.ts",
+    "test": "c8 tsx --env-file=.env.test --test --test-concurrency=1 test/api_tests/person.test.mjs test/api_tests/country.test.mjs test/api_tests/continents.test.mjs test/unit_tests/utils/query.test.mjs",
     "migrate": "knex migrate:latest",
     "seed": "knex seed:run"
   },
@@ -33,8 +36,13 @@
     "zod": "^3.25.76"
   },
   "devDependencies": {
+    "@types/express": "^5.0.6",
+    "@types/node": "^25.9.2",
+    "@types/supertest": "^7.2.0",
     "c8": "^9.1.0",
     "expect": "^29.7.0",
-    "supertest": "^6.3.4"
+    "supertest": "^6.3.4",
+    "tsx": "^4.22.4",
+    "typescript": "^6.0.3"
   }
 }
