+++ src/model/country.mjs
@@ -1,5 +1,5 @@
 import { BaseModel } from './base.mjs';
-import { applySort, applyFilter, applyPagination } from '../utils/query.mjs';
+import { applySort, applyFilter, applyPagination } from '../utils/query.js';
 
 /*
   * This class is responsible for handling the business logic of the country entity.
